添加了命令处理分支与R1 命令的处理程序

+++ CANDSPStringHandler/ControlStruct.h
@@ -3,7 +3,7 @@
 typedef enum ControlDataType ControlDataType;
 enum ControlDataType{
 	NONE,
-	INT16,
+	INT32,
 	FLOAT,
 };
